requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.5
django-allauth==0.44.0
django-countries==7.2.1
django-phonenumber-field==5.2.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
idna==2.10
oauthlib==3.1.1
phonenumbers==8.12.24
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.5
